+++ app/views/faqs/index.html.slim
@@ -3,7 +3,7 @@
     .faq_login
       = link_to(\
         t("shared.log_in"), \
-        new_auth_token_publishers_path, \
+        log_in_publishers_path, \
         class: "btn btn-block btn-secondary", \
         :"data-piwik-action" => "FAQLoginClicked", \
         :"data-piwik-name" => "Clicked", \
