+++ app/views/promo_registrations/_activated_unverified.html.slim
@@ -10,7 +10,7 @@
       - if current_publisher
         span = link_to(t("promo.activated.description_two"), home_publishers_path)
       - else
-        span = link_to(t("promo.activated.description_two"), new_auth_token_publishers_path)
+        span = link_to(t("promo.activated.description_two"), log_in_publishers_path)
   .promo--item-container
     span.promo--item-label= t("promo.shared.referral_links")
   p#promo--no-channels = t("promo.activated.unverified_no_channels")
@@ -18,5 +18,5 @@
     p.promo--body-text = t("promo.activated.unverified_error")
     - if current_publisher
       = link_to(t("promo.activated.unverified_cta").upcase, home_publishers_path, class: "btn btn-primary promo--activate-button")
-    - else 
-      = link_to(t("promo.activated.unverified_cta").upcase, new_auth_token_publishers_path, class: "btn btn-primary promo--activate-button")
+    - else
+      = link_to(t("promo.activated.unverified_cta").upcase, log_in_publishers_path, class: "btn btn-primary promo--activate-button")
