+++ app/views/promo_registrations/_active.html.slim
@@ -10,7 +10,7 @@
       - if current_publisher
         span = link_to(t("promo.active.description_three"), home_publishers_path)
       - else
-        span = link_to(t("promo.active.description_three"), new_auth_token_publishers_path)
+        span = link_to(t("promo.active.description_three"), log_in_publishers_path)
   .promo--item-container
     span.promo--item-label= t("promo.shared.referral_links")
   - n = 0
@@ -31,4 +31,4 @@
     - if current_publisher
       span = link_to(t("promo.active.description_three"), home_publishers_path)
     - else
-      span = link_to(t("promo.active.description_three"), new_auth_token_publishers_path)
+      span = link_to(t("promo.active.description_three"), log_in_publishers_path)
