+++ app/views/promo_registrations/_promo_footer.html.slim
@@ -8,6 +8,6 @@
       - if current_publisher
         = link_to(t("promo.shared.brave_rewards_tm"), home_publishers_path)
       - else
-        = link_to(t("promo.shared.brave_rewards_tm"), new_auth_token_publishers_path)
+        = link_to(t("promo.shared.brave_rewards_tm"), log_in_publishers_path)
     .footer--promo-item.footer--promo-item-tos
       = link_to(t("shared.terms_of_service"), terms_of_service_url)
