+++ app/views/publishers/registrations/emailed_auth_token.html.slim
