+++ app/views/publishers/registrations/expired_auth_token.html.slim
@@ -7,7 +7,7 @@
 
       .col-small-centered
         p= t ".body"
-        = form_for @publisher, {method: :post, url: create_auth_token_publishers_path} do |f|
-          = f.hidden_field(:email, :value => @publisher.email)
+        = form_for @publisher, {method: :post, url: registrations_path} do |f|
+          = hidden_field_tag(:email, @publisher.email)
           .panel-controls.col-center
             = f.submit(t(".button"), class: "btn btn-block btn-primary")
